+++ src/ydata_synthetic/preprocessing/regular/processor.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from typing import List, Optional
 
 from numpy import concatenate, ndarray, split, zeros
@@ -8,16 +7,6 @@ from sklearn.preprocessing import MinMaxScaler, OneHotEncoder
 from typeguard import typechecked
 
 from ydata_synthetic.preprocessing.base_processor import BaseProcessor
-
-
-class RegularModels(Enum):
-    "Supported models for the Regular Data Processor."
-    CGAN = 'CGAN'
-    CRAMERGAN = 'CramerGAN'
-    DRAGAN = 'DRAGAN'
-    GAN = 'VanillaGAN'
-    WGAN = 'WGAN'
-    WGAN_GP = 'WGAN_GP'
 
 @typechecked
 class RegularDataProcessor(BaseProcessor):
